WIP: blocktriple usage demo
