+++ education/internal/CMakeLists.txt
@@ -5,7 +5,7 @@ set(FP_SOURCES
     blockfraction.cpp
     blocksignificand.cpp
 )
-set(FP_SUBTYPE_SOURCES
+set(FP_ENGINE_SOURCES
     blocktriple.cpp
 )
 set(PERF_SOURCES
@@ -14,5 +14,5 @@ set(PERF_SOURCES
 
 compile_all("true" "demo" "Education/Internal/Multi-Limb/integer" "${INTEGER_SOURCES}")
 compile_all("true" "demo" "Education/Internal/Multi-Limb/floating-point" "${FP_SOURCES}")
-#compile_all("true" "demo" "Education/Internal/Multi-Limb/fp subtype" "${FP_SUBTYPE_SOURCES}")
+compile_all("true" "demo" "Education/Internal/Multi-Limb/fp-engine" "${FP_ENGINE_SOURCES}")
 #compile_all("true" "demo" "Education/Internal/Multi-Limb/performance" "${PERF_SOURCES}")
